# Temporary transactions page performance fix