This PR is a temporary solution to speed up the transactions page. A longer-term fix will fully replace this very hacky solution, but will require a larger rework of the search query.

Before:

| Type | IPS | Deviation | Time/Iteration | Iterations | Total Time |
|------|-----|-----------|----------------|------------|------------|
| COLD | 0.275 | ± 0.00%     | 3.64 s/i       | 1.000      | 3.638567s  |
| WARM | 0.431 | ± 0.00%     | 2.32 s/i       | 5.000      | 11.598782s |

After:

| Type | IPS | Deviation | Time/Iteration | Iterations | Total Time |
|------|-----|-----------|----------------|------------|------------|
| COLD | 0.344 | ± 0.00%     | 2.91 s/i       | 1.000      | 2.908487s  |
| WARM | 2.767 | ± 0.00%     | 361.45 ms/i    | 28.000     | 10.126426s |